% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMACS.R
\docType{data}
\name{BMACS}
\alias{BMACS}
\title{BMACS CD4 dataset}
\format{A data frame with 1817 rows and 6 variables.}
\usage{
data(BMACS)
}
\description{
This dataset is from the Baltimore site of the Multi-center AIDS Cohort Study (BMACS), which included 400 homosexual men who were infected by the human immunodeficiency virus (HIV) between 1984 and 1991.
}
\details{
\itemize{
  \item ID.      Subject ID
  \item Time.    Subject's study visit time
  \item Smoke.   Cigarette baseline smoking status
  \item age.     Age at study enrollment
  \item preCD4.  Pre-infection CD4 percentage
   \item CD4.     CD4 percentage at the time of visit
}
}
\references{
Kaslow, R. A., Ostrow, D. G., Detels, R., Phair, J. P., Polk, B. F. and Rinaldo, C. R. The Multicenter AIDS Cohort Study: rationale, organization and selected characteristics of the participants. American Journal of Epidemiology, 126:310-318, 1987.
}
\keyword{datasets}
