% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splinefit.R
\name{CVspline}
\alias{CVspline}
\title{Leave one-subject cross-validation score for spline fit}
\usage{
CVspline(Xvec, Yvec, ID, nKnots, Degree, Wt)
}
\arguments{
\item{Xvec, Yvec}{numeric vectors of data values, Xvec and Yvec must have the same length.}

\item{ID}{subject ID of the data value}

\item{nKnots}{number of equally-spaced knots}

\item{Degree}{degree of polynomial splines}

\item{Wt}{a weight vector. For longitudinal data, Wt=1/N corresponds to measurement uniform weight and Wt=1/(nni) corresponds subject uniform weight.}
}
\description{
Leave one-subject cross-validation score for spline fit
}
\references{
{ Wu, C.O. and Tian, X.   Nonparametric Models for Longitudinal Data. Chapman & Hall/CRC.
            To appear.}
}
