% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.model.fit.R
\name{DXi}
\alias{DXi}
\title{Derivative of the function Xi(s)}
\usage{
DXi(s)
}
\arguments{
\item{s}{a number or a vector}
}
\value{
value of the function DXi with give s
}
\description{
Derivative of the function Xi(s)
}
\examples{
DXi(c(-1000, -10,-5, 0, 5,10, 1000 ))
}
