% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel2D.R
\name{Kernel2D}
\alias{Kernel2D}
\title{2-dim Kernel function for longitudinal data}
\usage{
Kernel2D(IDls, Xvec, Yvec, X01, X02, Bndwdth1, Bndwdth2)
}
\arguments{
\item{IDls}{the vector of subject ID in a longitudinal sample}

\item{Xvec, }{Yvec  numeric vectors of data values, Xvec and Yvec must have the same length}

\item{X01, }{X02    two given values of Xvec}

\item{Bndwdth1, Bndwdth2}{two given bandwidths}
}
\value{
2-dim kernel fit result
}
\description{
2-dim Kernel function for longitudinal data
}
