% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel3D.R
\name{Kernel3D.S2}
\alias{Kernel3D.S2}
\title{3-dim Kernel function for longitudinal data to get  Pr(y2(t2)|x(t1))}
\usage{
Kernel3D.S2(IDls = ID, Y, Time, X, T1, T2, X0, Bndwdth1, Bndwdth2, Bndwdth3)
}
\arguments{
\item{IDls}{the vector of subject ID in a longitudinal sample}

\item{Y, X, Time}{numeric vectors of outcome, covariate and time of the the same length}

\item{T1, T2}{twp given time points}

\item{X0}{a given covariate value}

\item{Bndwdth1, Bndwdth2, Bndwdth3}{three bandwidths around two time and one covariate value}
}
\value{
3-dim Kernel function results
}
\description{
3-dim Kernel function for longitudinal data to get  Pr(y2(t2)|x(t1))
}
