% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{Kh.Ep}
\alias{Kh.Ep}
\title{Epanechnikov Kernel}
\usage{
Kh.Ep(datavec, Bndwdth)
}
\arguments{
\item{datavec}{a numeric vector}

\item{Bndwdth}{a bandwidth}
}
\value{
kernel function result
}
\description{
Epanechnikov Kernel
}
\examples{
Kh.Ep(2:7,5)
}
