% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalLm.Beta.R
\name{LocalLm.Beta.t0}
\alias{LocalLm.Beta.t0}
\title{Least square local linear fit at t0}
\usage{
LocalLm.Beta.t0(t0, Tvec, X1, X2, X3, Yvec, Bndwdth, Weight)
}
\arguments{
\item{t0}{a given time point}

\item{Tvec, Yvec}{numeric vectors of time and outcome values, Tvec and Yvec must have the same length.}

\item{X1, X2, X3}{three covariate vectors}

\item{Bndwdth}{a bandwidth of the Epanechnikov kernel}

\item{Weight}{the weight vector}
}
\description{
Least square local linear fit at t0
}
\examples{
 # see usage of LocalLm.Beta


}
