% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.model.fit.R
\name{Newton1var}
\alias{Newton1var}
\title{An equation solver with Newton's method with 1 variable}
\usage{
Newton1var(Z12vec, h0, Vh, HZB, Ind.Y, Diff = 1e-08, ORR, MaxIter = 100)
}
\arguments{
\item{MaxIter}{}
}
\value{
the root of the equation
}
\description{
An equation solver with Newton's method with 1 variable
}
