% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.model.fit.R
\name{Xi}
\alias{Xi}
\title{Function Xi(s)}
\usage{
Xi(s)
}
\arguments{
\item{s}{a number or a vector}
}
\value{
value of the function with give s
}
\description{
Function Xi(s)
}
\examples{
Xi(0)
Xi(c(-1000, -10,-5, 0, 5,10, 1000 ))
}
